refactor tests for pystiche.ops (#372)
